refactor: remove unused testing files for aggregatedpool module
