+++ registry/registry_test.go
